fix: handle types that can not be converted to string
There was a bug where if a value passed in was not convertible to a
string type, the library would throw

TypeError: Cannot convert object to primitive value

Any value that can not be converted to a string can not expose a
secret AFAICT so thus these types should just pass through without
issue. This patch implements that logic and adds tests with all object
types to validate that the function works as expected.